Move message context out of callback param
Actor message handlers now take only the actual messages as params
instead of including actor IDs and outbox too. Instead, these additional
fields are accessible on the mailbox or container for the duration of
handling a message batch. Note this pattern was implemented previously
for container outbox only.

The intent is to simplify the most common subscription use cases and
unify actor and container subscriptions.