Fix actor processing issue
When running tests with many threads, actor processing would eventually
stall. This was due to the use of TryEnter() in ActorProcessor, which
was intended to allow other threads to do other processing when
contention occurred. ActorProcessor.Enqueue() logic conditionally queues
actors and assumes the dispatcher will take full responsibility for
processing all messages, but the use of TryEnter() means that this
processing could be skipped in the case of contention.

One way contention could occur is: (1) worker has completed all messages
and has decided to release its lock, then (2) another message is
enqueued, causing a second worker to pick up the actor, then (3) the
second worker attempts to acquire the lock and fails, then (4) the first
worker releases the lock. The result is that neither worker will process
additional messages, causing the queue to remain non-empty and the actor
never enqueued to a worker again.

Switching to Enter() solves the immediate problem, but it could also
mean blocking for a variable amount of time when another thread is
processing messages. We could requeue the actor for later in such cases,
but then we need logic to handle backoffs and minimize spinning. Another
option is enqueueing to the worker currently processing the actor, but
this also adds complication.